temp

+++ populate_database.py
@@ -35,7 +35,7 @@ def generate_db_data_files(data_dir, tables, associations):
     i = 1
     table_name = a.name
     fp = open("{}/{}.tsv".format(fpath, table_name), 'w')
-    field_name = '|'.join(['id', '{}_id'.format(a.lft.name), '{}_id'.format(a.rgt.name)])
+    field_name = '|'.join(['id', a.assoc_f1, a.assoc_f2])
     fp.write('{}\n'.format(field_name))
     for j in range(1, actual_sizes[a.lft]+1):
       repeat_list = []
@@ -49,22 +49,25 @@ def generate_db_data_files(data_dir, tables, associations):
         print "finish {} lft".format(j)
     fp.close()
 
-def populate_database(data_dir, tables, associations, recreate_db=False):
-  if recreate_db:
-    remove_db()
-    create_db()
-    create_tables(tables, associations)
-  populate_tables(data_dir, tables, associations, recreate_db)
+def populate_database(data_dir, tables, associations, recreate=False):
+  remove_db(recreate)
+  create_db(recreate)
+  create_tables(tables, associations, recreate)
+  populate_tables(data_dir, tables, associations, recreate)
 
-def remove_db():
+def remove_db(recreate):
+  if recreate == False:
+    return
   db_name = get_db_name()
   sys_cmd("mysql -u root -e \"DROP DATABASE {};\"".format(db_name))
 
-def create_db():
+def create_db(recreate):
+  if recreate == False:
+    return
   db_name = get_db_name()
   sys_cmd("mysql -u root -e \"CREATE DATABASE {}\"".format(db_name))
 
-def create_tables(tables, associations):
+def create_tables(tables, associations, recreate):
   db_name = get_db_name()
   for t in tables:
     field_list=[]
@@ -73,10 +76,18 @@ def create_tables(tables, associations):
       sql_type = get_sql_type(f.tipe)
       ft += "{} DEFAULT NULL".format(sql_type)
       field_list.append(ft)
-    sys_cmd("mysql -u root {} -e \"CREATE TABLE {} ({});\"".format(db_name, get_db_table_name(t.name), ", ".join(field_list)))
+    if t.is_temp or recreate:
+      sys_cmd("mysql -u root {} -e \"CREATE TABLE {} ({});\"".format(db_name, get_db_table_name(t.name), ", ".join(field_list)))
+    if recreate == False:
+      temp_fields = []
+      for assoc in t.get_assocs():
+        if assoc.rgt == t and assoc.lft.is_temp:
+          f = '{}_id'.format(assoc.rgt_field_name)
+          sys_cmd("mysql -u root -e \"ALTER TABLE {} ADD COLUMN {} int(11) DEFAULT 0\"".format(to_plural(t.name), f))
   
-  for a in get_assoc_tables(associations):
-    sys_cmd("mysql -u root {} -e \"CREATE TABLE {} (id INTEGER, {}_id INTEGER, {}_id INTEGER);\"".format(db_name, a.name, a.lft.name, a.rgt.name))
+  if recreate:
+    for a in get_assoc_tables(associations):
+      sys_cmd("mysql -u root {} -e \"CREATE TABLE {} (id INTEGER, {} INTEGER, {} INTEGER);\"".format(db_name, a.name, a.assoc_f1, a.assoc_f2))
 
 def populate_tables(data_dir, tables, associations, recreate_db=False):
   db_name = get_db_name()
@@ -89,30 +100,27 @@ def populate_tables(data_dir, tables, associations, recreate_db=False):
         temp_fields.append('{}_id'.format(assoc.rgt_field_name))
     field_names = filter(lambda x: x != None, [None if f.name in temp_fields or f.is_temp else f.name for f in t.get_fields()])
     field_names = 'id,{}'.format(','.join(field_names))
-    for f in temp_fields:
-      if not recreate_db:
-        sys_cmd("mysql -u root -e \"ALTER TABLE {} ADD COLUMN {} int(11) DEFAULT 0\"".format(to_plural(t.name), f))
-    sys_cmd("mysql -u root --local-infile {} -e \"LOAD DATA LOCAL INFILE '{}' INTO TABLE {} FIELDS TERMINATED BY '|' LINES TERMINATED BY '\\n' IGNORE 1 LINES  ({})\"".format(db_name, fname, t.name, field_names))
+    sys_cmd("mysql -u root --local-infile {} -e \"LOAD DATA LOCAL INFILE '{}' INTO TABLE {} FIELDS TERMINATED BY '|' LINES TERMINATED BY '\\n' IGNORE 1 LINES  ({})\"".format(db_name, fname, to_plural(t.name), field_names))
     #sys_cmd('mysql {} -e "select * from {}" -B > {}/{}.tsv'.format(db_name, t.name, file_dir, t.name)) 
 
     #create index
     if recreate_db:
-      sys_cmd("mysql -u root {} -e \"CREATE INDEX index_{}_id ON {} (id) USING BTREE;\"".format(db_name, t.name, t.name))
+      sys_cmd("mysql -u root {} -e \"CREATE INDEX index_{}_id ON {} (id) USING BTREE;\"".format(db_name, t.name, to_plural(t.name)))
     
   for a in get_assoc_tables(associations):
     fname = "{}/{}.tsv".format(fpath, a.name)
-    field_names = ','.join(['id', '{}_id'.format(a.lft.name), '{}_id'.format(a.rgt.name)])
+    field_names = ','.join(['id', a.assoc_f1, a.assoc_f2])
     sys_cmd("mysql -u root --local-infile {} -e \"LOAD DATA LOCAL INFILE '{}' INTO TABLE {} FIELDS TERMINATED BY '|' LINES TERMINATED BY '\\n' IGNORE 1 LINES ({})\"".format(db_name, fname, a.name, field_names))
     #sys_cmd('mysql {} -e "select * from {}" -B > {}/{}.tsv'.format(db_name, table_name, file_dir, a.name))
 
     #create index
     #sys_cmd('mysql -u root {} -e "CREATE INDEX index_{}_id ON {} (id) USING BTREE;"'.format(db_name, table_name, table_name))
     if recreate_db:
-      sys_cmd('mysql -u root {} -e "CREATE INDEX index_{}_{}_id ON {} ({}_id) USING BTREE;"'.format(db_name, a.name, a.lft.name, a.name, a.lft.name))
-      sys_cmd('mysql -u root {} -e "CREATE INDEX index_{}_{}_id ON {} ({}_id) USING BTREE;"'.format(db_name, a.name, a.rgt.name, a.name, a.rgt.name))
+      sys_cmd('mysql -u root {} -e "CREATE INDEX index_{}_{}_id ON {} ({}) USING BTREE;"'.format(db_name, a.name, a.lft.name, a.name, a.assoc_f1))
+      sys_cmd('mysql -u root {} -e "CREATE INDEX index_{}_{}_id ON {} ({}) USING BTREE;"'.format(db_name, a.name, a.rgt.name, a.name, a.assoc_f2))
 
 
-def test_schema(tables, assertion=False):
+def test_schema(tables, associations):
   mysql_tables = []
   mysql_fields = {}
   mydb = mysql.connector.connect(host="localhost", user="root", passwd="", database=get_db_name())
@@ -126,13 +134,12 @@ def test_schema(tables, assertion=False):
     mysql_fields[t[0]] = [str(f[0]) for f in fields]
 
   for t in tables:
+    if t.is_temp:
+      continue
     db_table_name = get_db_table_name(t.name)
-    if assertion:
-      assert(db_table_name in mysql_tables, 'table {} not in mysql tables!'.format(db_table_name))
-    else:
-      if db_table_name not in mysql_tables:
-        print 'table {} not in mysql tables!'.format(db_table_name)
-        continue
+    if db_table_name not in mysql_tables:
+      print 'table {} not in mysql tables!'.format(db_table_name)
+      continue
     temp_fields = ['id']
     for assoc in t.get_assocs():
       if assoc.rgt == t and assoc.lft.is_temp:
@@ -140,9 +147,16 @@ def test_schema(tables, assertion=False):
     field_names = filter(lambda x: x != None, [None if f.name in temp_fields or f.is_temp else f.name for f in t.get_fields()])
     field_names.append('id')
     for f in field_names:
-      if assertion:
-        assert(f in mysql_fields[db_table_name], '  field {} not in mysql fields!'.format(f))
-      else:
-        if f not in mysql_fields[db_table_name]:
-          print '  field {} not in mysql fields of table {}!'.format(f, db_table_name)
+      if f not in mysql_fields[db_table_name]:
+        print '  field {} not in mysql fields of table {}!'.format(f, db_table_name)
+
+  for a in get_assoc_tables(associations):
+    field_names = ['id', a.assoc_f1, a.assoc_f2]
+    db_table_name = a.name
+    if db_table_name not in mysql_tables:
+      print 'table {} not in mysql tables!'.format(db_table_name)
+      continue
+    for f in field_names:
+      if f not in mysql_fields[db_table_name]:
+        print '  field {} not in mysql fields of table {}!'.format(f, db_table_name)
